Fix NumPy deprecations and clean gitignore
- replace deprecated np.float/np.int with modern types
- update subtype checks to np.integer
- remove merge artifact from .gitignore
- ignore data/ and zip outputs

+++ src/fluctmatch/commands/cmd_sca.py
@@ -76,7 +76,7 @@ from fluctmatch.analysis import (
     "--pcut",
     default=0.95,
     show_default=True,
-    type=np.float,
+    type=float,
     help="Cutoff value for sector selection")
 @click.option(
     "-r",
